// ntis-web/projects/itree-commons/src/lib/enums/table-row-actions.enums.ts
export enum ActionsEnum {
  ACTION_DISABLE_ORGANIZATION = 'DISABLE_ORGANIZATION',
  ACTION_INVITE_PERSON = 'INVITE_PERSON',
  ACTION_MANAGE_INTS_MANAGERS = 'MANAGE_INTS_MANAGERS',
  ACTION_REMOVE_FROM_ACCOUNT = 'REMOVE_FROM_ACCOUNT',
  ACTION_RENEW_INTS_INFO = 'RENEW_INTS_INFO',
  ACTION_VIEW_INTS_INFO = 'VIEW_INTS_INFO',
  ACTIONS_CANCEL = 'CANCEL',
  ACTIONS_CHANGE_TABLE_AUDIT = 'CHANGE_TABLE_AUDIT',
  ACTIONS_CONFIRM = 'CONFIRM',
  ACTIONS_COPY = 'COPY',
  ACTIONS_CREATE = 'CREATE',
  ACTIONS_DELETE = 'DELETE',
  ACTIONS_READ = 'READ',
  ACTIONS_REJECT = 'REJECT',
  ACTIONS_UPDATE = 'UPDATE',
  ADMIN_ACTIONS = 'ADMIN_ACTIONS',
  ASSIGN_ORG_ROLE = 'ASSIGN_ORG_ROLE',
  ASSIGN_ROLE = 'ASSIGN_ROLE',
  ASSIGN_USERS = 'ASSIGN_USERS',
  CAR_SPECIALIST_ACTIONS = 'CAR_SPECIALIST_ACTIONS',
  CREATE_OFFICIAL = 'CREATE_OFFICIAL',
  GLOBAL_FILTER_MANAGER = 'GLOBAL_FILTER_MANAGER',
  IMPORT_ORDERS = 'IMPORT_ORDERS',
  INTS_OWNER_ACTIONS = 'INTS_OWNER_ACTIONS',
  MANAGE_WF_PRIORITY_LIST = 'MANAGE_WF_PRIORITY_LIST',
  NEW_ORG_ACTIONS = 'NEW_ORG_ACTIONS',
  NTIS_ADMIN_ACTIONS = 'NTIS_ADMIN_ACTIONS',
  NTIS_ASSIGN_USERS = 'ASSIGN_USERS',
  ORDER_LAB_SERVICE = 'ORDER_LAB_SERVICE',
  PASL_ADMIN_ACTIONS = 'PASL_ADMIN_ACTIONS',
  READ_ALL = 'READ_ALL',
  REMOVE_ADMIN = 'REMOVE_ADMIN',
  SERVICE_PROVIDER_ACTIONS = 'SERVICE_PROVIDER_ACTIONS',
  SET_CODES_AND_TRANSLATIONS = 'SET_CODES_AND_TRANSLATIONS',
  SET_FORM_DISABLED_ACTIONS = 'SET_FORM_DISABLED_ACTIONS',
  SYSTEM_ADMIN_ACTIONS = 'SYSTEM_ADMIN_ACTIONS',
  TECH_SPECIALIST_ACTIONS = 'TECH_SPECIALIST_ACTIONS',
  UPDATE_STATE = 'UPDATE_STATE',
  UPLOAD_CONTRACT = 'UPLOAD_CONTRACT',
  VAND_ADMIN_ACTIONS = 'VAND_ADMIN_ACTIONS',
  WATER_MANAGER_ACTIONS = 'WATER_MANAGER_ACTIONS',
  NO_LIMIT_ON_ORG_LEVEL = 'NO_LIMIT_ON_ORG_LEVEL',
  GET_LINK = 'GET_LINK',
}